Spectrum viewer working, image viewer no longer displaying on spectrum viewer, not essential, fwhm and zero index working
