+++ .gitignore
@@ -8,3 +8,5 @@
 *.pyc
 
 .cursor/
+
+backup_files/
